Grille hexagonale

+++ src/main/java/engine/IGamePhysics.java
@@ -2,5 +2,5 @@ package engine;
 
 public interface IGamePhysics {
 
-    public void updatePhysics(float dt);
+    public void updatePhysics(double dt);
 }
