+++ src/main/java/model/CanadaPhysics.java
@@ -19,13 +19,13 @@ public class CanadaPhysics implements IGamePhysics {
     }
 
     @Override
-    public void updatePhysics(float dt) {
+    public void updatePhysics(double dt) {
 
         for (MovementComponent m : toUpdate){
             // On effectue les opérations à faire (collisions, deplacement, ...)
             GameObject gameObject = m.getGameObject();
-            float newX = gameObject.getX() + m.getVelocityX() * dt;
-            float newY = gameObject.getY() + m.getVelocityY() * dt;
+            double newX = gameObject.getX() + m.getVelocityX() * dt;
+            double newY = gameObject.getY() + m.getVelocityY() * dt;
             gameObject.setPosition(newX, newY);
 
             m.resetVelocity();
